Added forms/routes to create new user and recipes
Added a New User Registration form that checks against existing email addresses for uniqueness.

Consolidated a lot of html stubs and conditional logic for showing things such as session data into the base.html file. Used this to create a nav bar that shows certain nav items depending on whether the user is logged in and what page they're currently on.

Added a CSS reset for the sake of hopefully making styling easier and more consistent.

Removed the Type column from the Recipes table, since it's either too generic, might be what the Category attribute was supposed to be, or I just forgot why it's there. We'll see!

Also removed the Quantity column from Ingredients table, since it's probably not necessary considering it already exists in the Recipe_Ingredients table.

Added Breakfast/Lunch/Dinner Categories.

Added/edited comments for clarification.