Created New Recipe page
Created a new view_recipe route and page that the user is automatically forwarded to on successful completion of the Create Recipe function.

Restructured nav bar elements using Jinja2 conditionals to show when they're supposed to. Added Prev/Next recipe buttons based on the current page and recipe ids, allowing for endless clicking.

Added a 404/error page if the user tries to visit a recipe that doesn't exist.