Implemented Add Image to recipe function.
Added an add image function inside of the create_recipe page/route. Used Jinja 2 conditionals to show an image if the path is available.

Also set permitted filetypes for image uploads which should work...

Because this is a small app, the app will accept a file from the user's device and store it locally into an /images/ folder in the app.

Added error for users trying to access create_recipe while not logged in.

Updated .sql file to add the ImageURL attribute to the Recipe table.