Added styling, fixed images not showing on app start, other small fixes
Tried to make something that looks reasonably ok with some CSS. Previous and Next Recipe buttons were clarified and spaced out onto their own line. Added a favicon and logo. Made the logo/title a clickable item that takes the user home instead of a dedicated home link. Added a couple of recipes for testing. Fixed images not loading on app startup on the main page with the images route.