Updated allergens and dislikes, added aggregate function, minor fixes and CSS changes.
Split up allergen and dislike warnings and set to display the actual offending ingredient. Also fixed new user accounts with no allergens or dislikes specified throwing an error on the view_recipe pages.

Added a table to the User Settings page that shows all the user's recipes with links to click through. Also includes the results of an aggregate query to display the user's total number of recipes owned by them. This uses func.count from SQLAlchemy to calculate the total number of recipes owned by the user.

Removed CSS reset since we aren't using it.